doxygen: Corrected formatting.
Put the instruction on compilation flags as standard
comment instead of doxygen comment. I wanted the compliation
flags information to be in doxygen as part of the file description,

However, I cannot format
it correctly for doxygen. Before this step, the html output
is wrong, i.e. the description about compilation flags went
into the description of a function.

Signed-off-by: Ooi, Cinly <cinly.ooi@intel.com>