bugfix: Copy only valid data from heir_ids and hub_infos
Previously, will try to copy all elements from source array
to destination array , regardless of how many elements
available in the dstination  array.

The impact is, if we have more than 10 elements in the source
array, we have a buffer overflow in the destination array.
This is causing openocd to crash. The number 10 is the
default destination array size. It was chosen arbitarily.

The visible effect to user is, if the number of SLD nodes
is more than 10, then openocd will crash

Related:
hsdes:16018539775
aji_openocd:fed60fb998e26d3f4d2e30d46d6de071c1574494 -
    This is fixed together with this bug. The limit
    of 8 on this fix overwrites the 10 here

Signed-off-by: Ooi, Cinly <cinly.ooi@intel.com>